Chore: Refactor CreateChannelModal (#27469)
