+++ apps/meteor/client/sidebar/header/CreateChannel/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CreateChannelModal';
