+++ apps/meteor/client/sidebar/header/CreateTeam/CreateTeamModal.tsx
@@ -10,7 +10,7 @@ import { goToRoomById } from '../../../lib/utils/goToRoomById';
 
 type CreateTeamModalInputs = {
 	name: string;
-	description: string;
+	topic: string;
 	isPrivate: boolean;
 	readOnly: boolean;
 	encrypted: boolean;
@@ -93,7 +93,7 @@ const CreateTeamModal = ({ onClose }: { onClose: () => void }): ReactElement => 
 		members,
 		isPrivate,
 		readOnly,
-		description,
+		topic,
 		broadcast,
 		encrypted,
 	}: CreateTeamModalInputs): Promise<void> => {
@@ -104,7 +104,7 @@ const CreateTeamModal = ({ onClose }: { onClose: () => void }): ReactElement => 
 			room: {
 				readOnly,
 				extraData: {
-					description,
+					topic,
 					broadcast,
 					encrypted,
 				},
@@ -155,7 +155,7 @@ const CreateTeamModal = ({ onClose }: { onClose: () => void }): ReactElement => 
 							</Box>
 						</Field.Label>
 						<Field.Row>
-							<TextInput {...register('description')} placeholder={t('Teams_New_Description_Placeholder')} />
+							<TextInput {...register('topic')} placeholder={t('Teams_New_Description_Placeholder')} />
 						</Field.Row>
 					</Field>
 					<Field>
@@ -175,7 +175,7 @@ const CreateTeamModal = ({ onClose }: { onClose: () => void }): ReactElement => 
 							/>
 						</Box>
 					</Field>
-					<Field disabled={!canChangeReadOnly}>
+					<Field>
 						<Box display='flex' justifyContent='space-between' alignItems='start'>
 							<Box display='flex' flexDirection='column' width='full'>
 								<Field.Label>{t('Teams_New_Read_only_Label')}</Field.Label>
@@ -192,7 +192,7 @@ const CreateTeamModal = ({ onClose }: { onClose: () => void }): ReactElement => 
 							/>
 						</Box>
 					</Field>
-					<Field disabled={!canChangeEncrypted}>
+					<Field>
 						<Box display='flex' justifyContent='space-between' alignItems='start'>
 							<Box display='flex' flexDirection='column' width='full'>
 								<Field.Label>{t('Teams_New_Encrypted_Label')}</Field.Label>
