+++ apps/meteor/client/sidebar/header/actions/CreateRoomList.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 
 import CreateDiscussion from '../../../components/CreateDiscussion';
 import ListItem from '../../../components/Sidebar/ListItem';
-import CreateChannelWithData from '../CreateChannelWithData';
+import CreateChannelWithData from '../CreateChannel';
 import CreateDirectMessage from '../CreateDirectMessage';
 import CreateTeam from '../CreateTeam';
 import { useCreateRoomModal } from '../hooks/useCreateRoomModal';
