+++ apps/meteor/client/views/home/cards/CreateChannelsCard.tsx
@@ -4,7 +4,7 @@ import { useTranslation, useSetModal } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
 import React from 'react';
 
-import CreateChannelWithData from '../../../sidebar/header/CreateChannelWithData';
+import CreateChannelWithData from '../../../sidebar/header/CreateChannel';
 
 const CreateChannelsCard = (): ReactElement => {
 	const t = useTranslation();
