+++ apps/meteor/client/views/teams/contextualBar/channels/TeamsChannels.js
@@ -5,7 +5,7 @@ import React, { useCallback, useMemo, useState } from 'react';
 import { useRecordList } from '../../../../hooks/lists/useRecordList';
 import { AsyncStatePhase } from '../../../../lib/asyncState';
 import { roomCoordinator } from '../../../../lib/rooms/roomCoordinator';
-import CreateChannelWithData from '../../../../sidebar/header/CreateChannelWithData';
+import CreateChannelWithData from '../../../../sidebar/header/CreateChannel';
 import { useTabBarClose } from '../../../room/contexts/ToolboxContext';
 import RoomInfo from '../../../room/contextualBar/Info';
 import AddExistingModal from './AddExistingModal';
