+++ apps/meteor/packages/rocketchat-i18n/i18n/en.i18n.json
@@ -4094,6 +4094,7 @@
   "room_disallowed_reactions": "disallowed reactions",
   "Room_Edit": "Room Edit",
   "Room_has_been_archived": "Room has been archived",
+  "Room_has_been_created": "Room has been created",
   "Room_has_been_deleted": "Room has been deleted",
   "Room_has_been_removed": "Room has been removed",
   "Room_has_been_unarchived": "Room has been unarchived",
